Removed fixed column indexes and replaced with dynamic references.
Inserted variables and code to work with household numbering. Numbering works, but the new variable that should contain the household numbers produces an error when the name is dynamic (specified in the function input) instead of me typing the same information into that part of the function.

Have read Hadley's warning not to use paste0 as, and I quote, this is a ticking time bomb.

Have tried multiple methods to fix this, cannot locate a successful approach.

Error (using the TestWeight input at the bottom of the code):

Error in Create.SameSexCouples(HH3PPartneredMales, 0.1, TRUE, 0.6, 25,  :
  object 'MyHouseholds' not found