Start of opposite couple code. This includes a copy from the R script used for the 2-person households. This is the basis for the function.
Need to split out the separate input data frames/vectors. These need to be prepared by the user ahead of function use. For example, need a table of counts by age.

Current method assumes a skew normal distribution for the age differences. Need to decide if distributions for other age differences are required. If so, the script name will be amended to OppSexCouplesSN, with the other scripts including the relevant distribution used. Create.OppSexCouples will also need to be renamed to Create.OppSexCouples.SN for this function. The distribution will be used as the appended part of the function name, to keep things simple for naming conventions.