Activated the subfunction under the subfunctions part of the code and deleted the copy.
Started working on creating the pairs. Have reduced the number of iterations down dramatically just to get reasonably fast output so I can see if the code is working (this is being amended in the line calling the function, not inside the function, so I don't worry about forgetting!).

Will finish the pairing work tomorrow.

Some general code sectioning to make it clearer where blocks start and end, with an English explanation of the code. Should make it easier to maintain.