Added to the explanation of one parameter in the same sex couple function.
Have implemented Household numbering in the opposite sex function and is working perfectly. Have encountered the problem of trying to split out the Recipient and Donor dataframes based on the join appendix (.x or .y) plus the last column.

Rows 321 through 332 have showed the various attempted methods. If I don't get an error, I get code that runs but doesn't do the split.