More comments. Added in test of whether child is already matched to a same-sex school (*used later in school matching-rematching in multiple child families). Remaining children are ordered so that: 1. Children already matched to a same-sex school are ordered first. 2. Children who are not already matched to a same-sex school, but one is available, are next 3. Children who are unmatched to a same-sex school and cannot be matched to one, because there are no same-sex schools available for their age/sex combination are last.
The probabilities of being in a same-sex school are then reordered in descending order. This means that children who are already matched to a same-sex school have the highest probability of actually being matched to one.

This ordering means that the desired correlation (basically it is a correlation) entered by the user is matched as closely as possible by the function. If the reordering was not done, then there could be 0 matches: the ones with a same-sex match may be re-matched to a co-ed school because their prob was too low, and those matched to a co-ed school and there is no same-sex school available could be given a high prob. This would result in 0 matches to same-sex schools and must be avoided.