Have taken a back up of this file into the working branch.
Currently trying to fix out-of-range children w.r.t. age (here it's too old for the parent).

Current problem: have pulled a vector of household IDs that need to be ignored because they contain children with duplicate ages to that of the problem households (and therefore will incorrectly construct twins). BUT!!! The vector length is 61. These are 3-child households, so the number of children that should be ignored, at 3 person household is 61 * 3 = 183. And the total for the ones being not ignored is 174 (58 * 3). Why is there a 3-household (9-child) difference)?

Everything else is working. Sigh.