New function to create a non-parametric correlation between two variables, one of which must be a grouping variable and the other must be an ordinal variable.
Hours worked had stopped producing the correct output when a factor was ordered. The change to an ordered factor test broke the factor levels/labels at the end of the function. This has now been fixed so that the factor label test is only performed for an ordered factor (which is the only type of factor accepted, in addition to numeric variables).