Friends dataframe successfully created. Loop successful through dataframe. DataframeOfMatches is successfully constructed.
Only problem: every friend is getting their friend put into the DataframeofMatches. Sigh.

This is the next problem to fix. Only the friends of one friend (and, in this total looping method, the last friend in the dataframe) should be in the DataframeOfMatches. Will be because I haven't deleted the object within the "goes through the friends dataframe" loop.