Working through so that there is a new sampled parent for the ones that are too old or too young for their children.
Currently, the out-of-age parent has been removed from the working parent data frame.

Next test is to extract the problem child. The function will attempt to swap the problem child to another household (with a reverse swap to preserve the numbers) so that the parental ages remain correct given the input parameters.