I think there are some problems that originated in how I handled the extraction of the extra children. I've now changed it. Any mismatch from this point on, trying to get the original children back, is due to sampling.
Note: the new approach is to pull all x+2:n out into a long file, with only the households and ages. There will be a counter put against each age for the dataframe to match, and this long file, so that age + counter is matched.

First I need to test that the same number of children at each age is present in the dataframes that will be matched. But at least I got to this damn point.