Commented out debugging prints to console. Now working on fixing the households that contain twins but shouldn't. Problem household correctly identified.
Now to work the same as fixing the parent ages EXCEPT that a simple household ID swap cannot be used. One of the twins must be swapped for another of a different same age, from a different household, so that twins are not constructed in the swap household (otherwise not fixing the problem).