Deleted data frames that must only exist within each household and not be available for the next household.
Now have a summary of schools by the number of children that can go into them and the associated roll counts.

Have created a merger of the same sex schools so that if:
1. at least two children must go to the same school, AND
2. there are children of the opposite sex who could go to same-sex schools in the same age range on the rolls AND
3. at least two children must go to the same school

THEN
the rolls for the ages for the same-sex schools are taken into account.

STILL MAY NEED FIXING
not sure about the summing of the roll counts. Is done by age. That is, if a same-sex boys' school can take a person of one age and the same sex girls' school can take a girl of the same age range, BUT NOT THE SAME AGE, how do I sort this out?

STILL THINKING.