Have got the counts done for the co-ed schools. Now trying to work out how to do them for the same-sex schools.
Problem = must sum the rolls for the two sexes, for the relevant ages. E.g. if a male aged 15 at Boys' A and a girl aged 16 at Girls's A then the roll for the 16-year-olds at the former and the roll for the 15-year-olds at the latter must be constructed.

BUT: need to ensure that the minimum required numbers exist. So a little stuck.