Fix an error, get another one.
Fixed: when there is a female school with no matching male one (was producing NA results and should not have tried to match an empty dataframe)

Added single-sex schools where there is no matched single-sex school, to the list of schools from which to sample.

New error: paste0 line, error is "SchoolID must be size 0 or 1, not 2"
About to debug this.