Joins and then summarise, argh. I thought I had redone the SchoolsSubset successfully with the PossibleSchools, but no.
The problem was the join used.

Note: for this, an inner_join returns the same matches as a left_join, but left_join is easier to understand. The by is giving 0s instead of NAs when there is no age match in SchoolsRenamed, hence filtering by 0 and not NA.

Bug in join?