Working through the code with the first household that gets selected. So far, so good. There are no matching single-sex schools for this household (girl who may go to either coed or female college, twins who are going to the same primary school).
Primary school correctly selected (takes 2 kids, not 1, so has priority in selection). Now have to actually put these kids into the same school.

Then: correctly add a SINGLE school to the girl. This was a bit that was falling over the other day. Hopefully fixed with the changes I have made to the code.