Okay, so far have fixed: 1. selection of school based on the number of kids for each age/school/sex or coed combination 2. putting the right school against the right child 3. getting the roll count correctly decreased for making sure that the child can go to the school
Problem
Because the roll count decreases are by age within the school, if there are three kids going to the same school, but different ages, the maximum number of kids the school can take is the number of kids the same age

Solution?
Create a separate counter data frame that is the number of kids per school. Test the school choice against this counter, but retain the school roll counts in the original SchoolSubset file.