Still reworking the code so that multiple children, of different ages, at the same school are selected on 1. the number of children at the same school and then 2. probability of school selection based on relative counts of children for all ages that will be at the same school
Seems to be working. Will need rework when I come across a household where there are children of both sex who can both go to a single-sex school. Trying to get the uncomplicated code working first before I deal with that amendment. That said, the code is working through the first two households.