The combo single-sex schools now correctly selected and split into the separate single-sex schools. The next objective is to ensure that only the correct number of children are allocated.
Problem to solve (example):
There are three children in two single-sex schools. Both schools are selected from the combo. Only two children can go to the same school.

Thus: the match needs to be made on the repeated same-sex school. And the other same-sex school rejected for both the current children and the other children in the family.