Fixed the problem with a missing else.
New problem:
Situation: school contains X number of places
Requirement: school must take X-N children

Current outcome: code falls over with filter not working on lines 690-691, with error:
Error: Problem with `filter()` input `..1`.
ℹ Input `..1` is `ChildAge == AgeToUse$ChildAge`.
x Input `..1` must be of size 3 or 1, not size 0.
Run `rlang::last_error()` to see where the error occurred.
In addition: Warning message:
In max(CheckForMultiples$NumberKidsThatAge) :
  no non-missing arguments to max; returning -Inf

Problem household is 687. This is the first household where only 2 of 3 children can go to the same school, but the children are all different ages. In the previous households with, 2 of 3 children needing to go to the same school, 2 of the children were the same age.