Numbers are coming out right just ahead of new bit.
New problem: minimum age in parent dataset may be older than minimum age stated in input.

Need to add in the difference between min age specified and max age permitted so that index works correctly. Currently, index is out by 1 for getting the correct age. Grabs an age, decreases correctly, but it is the wrong age.