Fixed out of range minimum index value (i.e. min parent age in the parent data frame is older than the min parent age allowed, only happens for very young children).
Removed code no longer used to match last lot of kids.

Constructed two data frames from last lot of kids:
Kids that can be matched to a parent as a parent exists for the required age range, this is appended to the kids already matched

Kids with no matching parent, will often be an empty data frame. Needs testing.

Next job: output as a list.