Redoing the childrenyes into childrenno
Have a temp household ID added to the parents df at the start. This will have holes at the end as not all parents will match, and they will match out of order.

At the end, need to redo the household numbers so that they reflect the numbers provided by the user.