Starting the cohabitating with no pre-existing household function.
I have thought through sex-segregation and have concluded that the only way forward is completely random or completely sex-segregated.

This is because proportions between 0 and 1 don't make practical sense. In a completely random assignment, some households will be randomly same-sex. The probability of this occurring will decrease with increasing household size.