Split out the sexes. If the sex dataframe is not modulo 0 for household size, then a random number of people equal to the modulo result is removed.
Split between base and people to add completed.

NOTE: this is all for the case when the houses must have complete sex segregation. Will be the same for those without sex segregation as the only difference in logic here is the split out of the separate sexes.

Not sure if this approach is the way to go. Asking supervisors, but will continue with the programming as this part is standard irrespective of which situation (same-sex only or opposite sex).