It's working now, changing the test to whether something is numeric.
While the employer id will be set to either 0 or "0", all other data from the employer file will be NA. As I have no idea what is in the employer file, other than the minimum required for the assignment, I can't easily set all employer-extracted columns to a value such as 0. This is a user problem to deal with.