Updated the initial socnet code so it calls for a matrix of integers - this is the network. Added in two tests for it.
Removed the network variable from the Networks data frame, as this must now be supplied from a separate object. That will remove all expectation that the network counts used relate to anyone in the data frame until the output.

Created a matrix of networks, and added it to the data in the package.