Deleted parts that related to using a network column in a data frame
Test is now whether the sum of the matrix is even. If not, a random position (index) is drawn from the matrix and 1 is added.

Information message printed to console:
That networks links are odd
That 1 has been added at the named position
That the value has been incremented from original to original+1