Updated code examples to include a smaller example, which will be used in the article.
Consequential changes to the setting up data script.

Added in examples to the socnet function.

Updated the matrix data to reflect the fact that the new 5K sample is a reproducible grab from Township.

Updated the associated matrix data so it had the correct length.

Added in new 5K sample file from Township. Changed the name as "Networks" may be a protected name somewhere.