Interdiff function amended in ThesisChapter and copied over. Removed the seed as it's not used in the function. Fixed the grouping variable name internal to the function, was using the dataframe name provided to the function. Now using the internal name.
BadRels is the data frame with the relationship data that is the outcome of the age assignment. It is the input for the interdiff function and is part of the provided package data frames.

GroupInfo will probably get deleted as I think I can do it in the code example. Keeping it in for the moment, but expecting to delete it. The format is incorrect anyway, it's not one line per group. Either way, it's going to be heavily amended or deleted. At least its here for posperity.