Redone the Township set-up so all datasets need to be reworked due to change in seed value.
relstatfix will also do other variables, depending on group definition and value.

The code examples and setting up scripts changed so that old now irrelevant work is commented out and new work is added in. E.g. the graphs for the redistribution of school kids' working hours. The extra graph and correlation, and so forth.